schema: add empty DB upgrade path from 4.14.0.0 to 4.15.0.0 (#4092)
engine/schema: add empty DB upgrade path from 4.14.0.0 to 4.15.0.0
